revert change to transaction wrapper
