+++ back/src/bsds/resolvers/queries/__tests__/bsds.bsdd.integration.ts
@@ -20,7 +20,6 @@ import {
   formFactory,
   toIntermediaryCompany
 } from "../../../../__tests__/factories";
-import prisma from "../../../../prisma";
 
 import { indexForm } from "../../../../forms/elastic";
 import { getFullForm } from "../../../../forms/database";
@@ -161,15 +160,6 @@ describe("Query.bsds workflow", () => {
         }
       );
       formId = id;
-      const form = await prisma.form.findUnique({
-        where: { id: formId },
-        include: {
-          transportSegments: true,
-          intermediaries: true,
-          forwardedIn: true
-        }
-      });
-      await indexForm(form);
       await refreshElasticSearch();
     });
 
@@ -239,16 +229,6 @@ describe("Query.bsds workflow", () => {
           }
         }
       );
-      const form = await prisma.form.findUnique({
-        where: { id: formId },
-        include: {
-          transportSegments: true,
-          intermediaries: true,
-          forwardedIn: true
-        }
-      });
-
-      await indexForm(form);
       await refreshElasticSearch();
     });
 
@@ -316,15 +296,6 @@ describe("Query.bsds workflow", () => {
           }
         }
       });
-      const form = await prisma.form.findUnique({
-        where: { id: formId },
-        include: {
-          transportSegments: true,
-          intermediaries: true,
-          forwardedIn: true
-        }
-      });
-      await indexForm(form);
       await refreshElasticSearch();
     });
 
@@ -391,15 +362,6 @@ describe("Query.bsds workflow", () => {
           }
         }
       });
-      const form = await prisma.form.findUnique({
-        where: { id: formId },
-        include: {
-          transportSegments: true,
-          intermediaries: true,
-          forwardedIn: true
-        }
-      });
-      await indexForm(form);
       await refreshElasticSearch();
 
       expect(errors).toBeUndefined();
@@ -448,15 +410,6 @@ describe("Query.bsds workflow", () => {
           }
         }
       });
-      const form = await prisma.form.findUnique({
-        where: { id: formId },
-        include: {
-          transportSegments: true,
-          intermediaries: true,
-          forwardedIn: true
-        }
-      });
-      await indexForm(form);
       await refreshElasticSearch();
 
       expect(errors).toBeUndefined();
