+++ back/src/forms/resolvers/mutations/markAsAccepted.ts
@@ -11,6 +11,7 @@ import transitionForm from "../../workflow/transitionForm";
 import { EventType } from "../../workflow/types";
 import { renderFormRefusedEmail } from "../../mail/renderFormRefusedEmail";
 import { sendMail } from "../../../mailer/mailing";
+import { runInTransaction } from "../../../common/repository/helper";
 
 const markAsAcceptedResolver: MutationResolvers["markAsAccepted"] = async (
   _,
@@ -42,7 +43,7 @@ const markAsAcceptedResolver: MutationResolvers["markAsAccepted"] = async (
         signedAt: new Date(acceptedInfo.signedAt)
       };
 
-  const acceptedForm = await prisma.$transaction(async transaction => {
+  const acceptedForm = await runInTransaction(async transaction => {
     const { update, removeAppendix2 } = getFormRepository(user, transaction);
     const acceptedForm = await update(
       { id: form.id },
