+++ back/src/forms/resolvers/mutations/markAsReceived.ts
@@ -18,6 +18,7 @@ import {
 } from "@prisma/client";
 import { renderFormRefusedEmail } from "../../mail/renderFormRefusedEmail";
 import { sendMail } from "../../../mailer/mailing";
+import { runInTransaction } from "../../../common/repository/helper";
 
 const markAsReceivedResolver: MutationResolvers["markAsReceived"] = async (
   parent,
@@ -67,7 +68,7 @@ const markAsReceivedResolver: MutationResolvers["markAsReceived"] = async (
         currentTransporterSiret: ""
       };
 
-  const receivedForm = await prisma.$transaction(async transaction => {
+  const receivedForm = await runInTransaction(async transaction => {
     const formRepository = getFormRepository(user, transaction);
     const receivedForm = await formRepository.update(
       { id: form.id },
