+++ back/src/forms/resolvers/mutations/markAsTempStored.ts
@@ -12,6 +12,7 @@ import { getFormRepository } from "../../repository";
 import prisma from "../../../prisma";
 import { renderFormRefusedEmail } from "../../mail/renderFormRefusedEmail";
 import { sendMail } from "../../../mailer/mailing";
+import { runInTransaction } from "../../../common/repository/helper";
 
 const markAsTempStoredResolver: MutationResolvers["markAsTempStored"] = async (
   parent,
@@ -54,7 +55,7 @@ const markAsTempStoredResolver: MutationResolvers["markAsTempStored"] = async (
       : {})
   };
 
-  const tempStoredForm = await prisma.$transaction(async transaction => {
+  const tempStoredForm = await runInTransaction(async transaction => {
     const formRepository = getFormRepository(user, transaction);
     const tempStoredForm = await formRepository.update(
       { id: form.id },
