+++ back/src/forms/resolvers/mutations/markAsTempStorerAccepted.ts
@@ -11,6 +11,7 @@ import { acceptedInfoSchema } from "../../validation";
 import prisma from "../../../prisma";
 import { renderFormRefusedEmail } from "../../mail/renderFormRefusedEmail";
 import { sendMail } from "../../../mailer/mailing";
+import { runInTransaction } from "../../../common/repository/helper";
 
 const markAsTempStorerAcceptedResolver: MutationResolvers["markAsTempStorerAccepted"] =
   async (_, args, context) => {
@@ -38,7 +39,7 @@ const markAsTempStorerAcceptedResolver: MutationResolvers["markAsTempStorerAccep
       }
     };
 
-    const tempStoredForm = await prisma.$transaction(async transaction => {
+    const tempStoredForm = await runInTransaction(async transaction => {
       const formRepository = getFormRepository(user, transaction);
       const tempStoredForm = await formRepository.update(
         { id: form.id },
